+++ libraries/chaos/include/chaos/SpriteManager.h
@@ -60,7 +60,7 @@ namespace chaos
     /** Display the sprites */
     void Display(GLProgramVariableProvider * uniform_provider);
     /** get the number of sprites in the buffer */
-    int GetSpriteCount() const { return sprites.size() / 6; }
+    size_t GetSpriteCount() const { return sprites.size() / 6; }
     /** remove the sprites (GPU buffer will be clean at next UpdateGPUVertexBuffer)*/
     void ClearSprites();
     /** get the program */
